with crash + mini fixes

+++ crates/lgn-graphics-renderer/src/gpu_renderer/render_layer.rs
@@ -21,7 +21,7 @@ pub struct RenderLayer {
 impl RenderLayer {
     pub fn new(allocator: &UnifiedStaticBufferAllocator, cpu_render_set: bool) -> Self {
         const TEMP_MAX_MATERIAL_COUNT: usize = 8192;
-        let page_size = TEMP_MAX_MATERIAL_COUNT * std::mem::size_of::<u32>();
+        let page_size = TEMP_MAX_MATERIAL_COUNT * std::mem::size_of::<u64>();
         let state_page = allocator.allocate(page_size as u64, ResourceUsage::empty());
 
         Self {
